[feat] make glue job dag for followers
